Use two files for UploadButton
