+++ js/uploadbutton/Index.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
 	import type { Gradio } from "@gradio/utils";
-	import { tick } from "svelte";
-	import { BaseButton } from "@gradio/button";
 	import type { FileData } from "@gradio/upload";
-	import { upload, prepareFiles } from "@gradio/upload";
+	import UploadButton from "./shared/UploadButton.svelte";
 
 	export let elem_id = "";
 	export let elem_classes: string[] = [];
@@ -25,85 +23,31 @@
 	export let mode: "static" | "interactive";
 
 	$: disabled = mode === "static";
-
-	let hidden_upload: HTMLInputElement;
-	let accept_file_types: string | null;
-
-	if (file_types == null) {
-		accept_file_types = null;
-	} else {
-		file_types = file_types.map((x) => {
-			if (x.startsWith(".")) {
-				return x;
-			}
-			return x + "/*";
-		});
-		accept_file_types = file_types.join(", ");
-	}
-
-	function openFileUpload(): void {
-		hidden_upload.click();
-	}
-
-	async function loadFiles(files: FileList): Promise<void> {
-		let _files: File[] = Array.from(files);
-		if (!files.length) {
-			return;
-		}
-		if (file_count === "single") {
-			_files = [files[0]];
-		}
-		let all_file_data = await prepareFiles(_files);
-		await tick();
-		all_file_data = await upload(all_file_data, root);
-		console.log(all_file_data);
-		gradio.dispatch("change", all_file_data);
-		gradio.dispatch("upload", all_file_data);
-		value = all_file_data;
-	}
-
-	async function loadFilesFromUpload(e: Event): Promise<void> {
-		const target = e.target as HTMLInputElement;
-
-		if (!target.files) return;
-		await loadFiles(target.files);
-	}
-
-	function clearInputValue(e: Event): void {
-		const target = e.target as HTMLInputElement;
-		if (target.value) target.value = "";
-	}
 </script>
 
-<input
-	class="hide"
-	accept={accept_file_types}
-	type="file"
-	bind:this={hidden_upload}
-	on:change={loadFilesFromUpload}
-	on:click={clearInputValue}
-	multiple={file_count === "multiple" || undefined}
-	webkitdirectory={file_count === "directory" || undefined}
-	mozdirectory={file_count === "directory" || undefined}
-	data-testid="{label}-upload-button"
-/>
-
-<BaseButton
-	{size}
-	{variant}
+<UploadButton
 	{elem_id}
 	{elem_classes}
 	{visible}
-	on:click={openFileUpload}
+	{file_count}
+	{file_types}
+	{size}
 	{scale}
 	{min_width}
+	{root}
+	{value}
 	{disabled}
+	{variant}
+	{label}
+	on:click={() => gradio.dispatch("click")}
+	on:change={({ detail }) => {
+		value = detail;
+		gradio.dispatch("change", detail);
+	}}
+	on:upload={({ detail }) => {
+		value = detail;
+		gradio.dispatch("upload", detail);
+	}}
 >
 	{gradio.i18n(label)}
-</BaseButton>
-
-<style>
-	.hide {
-		display: none;
-	}
-</style>
+</UploadButton>
